+++ android-components/components/feature/accounts/src/main/java/mozilla/components/feature/accounts/FirefoxAccountsAuthFeature.kt
@@ -11,6 +11,7 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.launch
 import mozilla.components.concept.engine.EngineSession
 import mozilla.components.concept.engine.request.RequestInterceptor
+import mozilla.components.concept.sync.FxAEntrypoint
 import mozilla.components.service.fxa.FxaAuthData
 import mozilla.components.service.fxa.manager.FxaAccountManager
 import mozilla.components.service.fxa.toAuthType
@@ -32,15 +33,15 @@ class FirefoxAccountsAuthFeature(
     private val coroutineContext: CoroutineContext = Dispatchers.IO,
     private val onBeginAuthentication: (Context, String) -> Unit = { _, _ -> },
 ) {
-    fun beginAuthentication(context: Context) {
+    fun beginAuthentication(context: Context, entrypoint: FxAEntrypoint) {
         beginAuthenticationAsync(context) {
-            accountManager.beginAuthentication()
+            accountManager.beginAuthentication(entrypoint = entrypoint)
         }
     }
 
-    fun beginPairingAuthentication(context: Context, pairingUrl: String) {
+    fun beginPairingAuthentication(context: Context, pairingUrl: String, entrypoint: FxAEntrypoint) {
         beginAuthenticationAsync(context) {
-            accountManager.beginAuthentication(pairingUrl)
+            accountManager.beginAuthentication(pairingUrl, entrypoint = entrypoint)
         }
     }