+++ android-components/components/service/firefox-accounts/src/main/java/mozilla/components/service/fxa/FirefoxAccount.kt
@@ -13,6 +13,7 @@ import kotlinx.coroutines.withContext
 import mozilla.components.concept.base.crash.CrashReporting
 import mozilla.components.concept.sync.AuthFlowUrl
 import mozilla.components.concept.sync.DeviceConstellation
+import mozilla.components.concept.sync.FxAEntrypoint
 import mozilla.components.concept.sync.MigratingAccountInfo
 import mozilla.components.concept.sync.OAuthAccount
 import mozilla.components.concept.sync.StatePersistenceCallback
@@ -94,9 +95,9 @@ class FirefoxAccount internal constructor(
         persistCallback.setCallback(callback)
     }
 
-    override suspend fun beginOAuthFlow(scopes: Set<String>, entryPoint: String) = withContext(scope.coroutineContext) {
+    override suspend fun beginOAuthFlow(scopes: Set<String>, entryPoint: FxAEntrypoint) = withContext(scope.coroutineContext) {
         handleFxaExceptions(logger, "begin oauth flow", { null }) {
-            val url = inner.beginOAuthFlow(scopes.toTypedArray(), entryPoint)
+            val url = inner.beginOAuthFlow(scopes.toTypedArray(), entryPoint.entrypointName)
             val state = Uri.parse(url).getQueryParameter("state")!!
             AuthFlowUrl(state, url)
         }
@@ -105,13 +106,13 @@ class FirefoxAccount internal constructor(
     override suspend fun beginPairingFlow(
         pairingUrl: String,
         scopes: Set<String>,
-        entryPoint: String,
+        entryPoint: FxAEntrypoint,
     ) = withContext(scope.coroutineContext) {
         // Eventually we should specify this as a param here, but for now, let's
         // use a generic value (it's used only for server-side telemetry, so the
         // actual value doesn't matter much)
         handleFxaExceptions(logger, "begin oauth pairing flow", { null }) {
-            val url = inner.beginPairingFlow(pairingUrl, scopes.toTypedArray(), entryPoint)
+            val url = inner.beginPairingFlow(pairingUrl, scopes.toTypedArray(), entryPoint.entrypointName)
             val state = Uri.parse(url).getQueryParameter("state")!!
             AuthFlowUrl(state, url)
         }